Adding tests to store_config. In-Memory stores seem to be jacked though
